ls-refs: introduce ls-refs server command
Introduce the ls-refs server command.  In protocol v2, the ls-refs
command is used to request the ref advertisement from the server.  Since
it is a command which can be requested (as opposed to mandatory in v1),
a client can sent a number of parameters in its request to limit the ref
advertisement based on provided ref-prefixes.

Change-Id: Ib9df7f193588ac8b7164b27b005c5ed4db7042ae
Signed-off-by: Brandon Williams <bmwill@google.com>