transport: store protocol version
Once protocol_v2 is introduced requesting a fetch or a push will need to
be handled differently depending on the protocol version.  Store the
protocol version the server is speaking in 'struct git_transport_data'
and use it to determine what to do in the case of a fetch or a push.

Change-Id: If4df7dbb97ff0773290cf3e13fc5176c4bf78e43
Signed-off-by: Brandon Williams <bmwill@google.com>