remote-curl: implement stateless-connect command
Teach remote-curl the 'stateless-connect' command which is used to
establish a stateless connection with servers which support protocol
version 2.  This allows remote-curl to act as a proxy, allowing the git
client to communicate natively with a remote end, simply using
remote-curl as a pass through to convert requests to http.

Change-Id: I7af7adc48b6fa733d231413d0666d8e5a11927c1
Signed-off-by: Brandon Williams <bmwill@google.com>