refspec: introduce struct refspec
Introduce 'struct refspec', an abstraction around a collection of
'struct refspec_item's much like how 'struct pathspec' holds a
collection of 'struct pathspec_item's.

A refspec struct also contains an array of the original refspec strings
which will be used to facilitate the migration to using this new
abstraction throughout the code base.

Change-Id: I84a23e8fc90979808514c178fe655606d8909be9
Signed-off-by: Brandon Williams <bmwill@google.com>