fetch: refactor to make function args narrower
Refactor find_non_local_tags and get_ref_map to only take the
information they need instead of the entire transport struct. Besides
improving code clarity, this also improves their flexibility, allowing
for a different set of refs to be used instead of relying on the ones
stored in the transport struct.

Change-Id: I9c7ab54b0e6980eaccd936bf2df914d19cc76ffc
Signed-off-by: Brandon Williams <bmwill@google.com>